enabled hist method for cpu; added support of unified descriptor for infer and train (#1445)
enabled hist method for cpu; added support of unified descriptor for df infer and train
