+++ cpp/oneapi/dal/algo/decision_forest/backend/cpu/infer_kernel_cls.cpp
@@ -112,8 +112,8 @@ static result_t infer(const context_cpu& ctx, const descriptor_t& desc, const in
     return call_daal_kernel<Float>(ctx, desc, input.get_model(), input.get_data());
 }
 
-template <typename Float, typename Task>
-struct infer_kernel_cpu<Float, Task, method::dense> {
+template <typename Float>
+struct infer_kernel_cpu<Float, method::by_default, task::classification> {
     result_t operator()(const context_cpu& ctx,
                         const descriptor_t& desc,
                         const input_t& input) const {
@@ -121,7 +121,7 @@ struct infer_kernel_cpu<Float, Task, method::dense> {
     }
 };
 
-template struct infer_kernel_cpu<float, task::classification, method::dense>;
-template struct infer_kernel_cpu<double, task::classification, method::dense>;
+template struct infer_kernel_cpu<float, method::by_default, task::classification>;
+template struct infer_kernel_cpu<double, method::by_default, task::classification>;
 
 } // namespace oneapi::dal::decision_forest::backend
