+++ cpp/oneapi/dal/algo/decision_forest/backend/gpu/infer_kernel.hpp
@@ -21,7 +21,7 @@
 
 namespace oneapi::dal::decision_forest::backend {
 
-template <typename Float, typename Task, typename Method>
+template <typename Float, typename Method, typename Task>
 struct infer_kernel_gpu {
     infer_result<Task> operator()(const dal::backend::context_gpu& ctx,
                                   const detail::descriptor_base<Task>& params,
