+++ cpp/oneapi/dal/algo/decision_forest/backend/gpu/train_kernel.hpp
@@ -21,7 +21,7 @@
 
 namespace oneapi::dal::decision_forest::backend {
 
-template <typename Float, typename Task, typename Method>
+template <typename Float, typename Method, typename Task>
 struct train_kernel_gpu {
     train_result<Task> operator()(const dal::backend::context_gpu& ctx,
                                   const detail::descriptor_base<Task>& params,
