+++ cpp/oneapi/dal/algo/decision_forest/backend/gpu/train_kernel_cls_hist_dpc.cpp
@@ -171,7 +171,7 @@ static result_t train(const context_gpu& ctx, const descriptor_t& desc, const in
 }
 
 template <typename Float, typename Task>
-struct train_kernel_gpu<Float, Task, method::hist> {
+struct train_kernel_gpu<Float, method::hist, Task> {
     result_t operator()(const context_gpu& ctx,
                         const descriptor_t& desc,
                         const input_t& input) const {
@@ -179,7 +179,7 @@ struct train_kernel_gpu<Float, Task, method::hist> {
     }
 };
 
-template struct train_kernel_gpu<float, task::classification, method::hist>;
-template struct train_kernel_gpu<double, task::classification, method::hist>;
+template struct train_kernel_gpu<float, method::hist, task::classification>;
+template struct train_kernel_gpu<double, method::hist, task::classification>;
 
 } // namespace oneapi::dal::decision_forest::backend
