+++ cpp/oneapi/dal/algo/decision_forest/detail/infer_ops.hpp
@@ -33,7 +33,7 @@ template <typename Descriptor>
 struct infer_ops {
     using float_t = typename Descriptor::float_t;
     using task_t = typename Descriptor::task_t;
-    using method_t = typename Descriptor::method_t;
+    using method_t = method::by_default;
     using input_t = infer_input<task_t>;
     using result_t = infer_result<task_t>;
     using descriptor_base_t = descriptor_base<task_t>;
