+++ cpp/oneapi/dal/algo/decision_forest/detail/train_ops.cpp
@@ -29,7 +29,7 @@ struct train_ops_dispatcher<host_policy, Float, Task, Method> {
                                   const descriptor_base<Task>& desc,
                                   const train_input<Task>& input) const {
         using kernel_dispatcher_t =
-            dal::backend::kernel_dispatcher<backend::train_kernel_cpu<Float, Task, Method>>;
+            dal::backend::kernel_dispatcher<backend::train_kernel_cpu<Float, Method, Task>>;
         return kernel_dispatcher_t()(ctx, desc, input);
     }
 };
