+++ cpp/oneapi/dal/algo/decision_forest/detail/train_ops_dpc.cpp
@@ -30,8 +30,8 @@ struct train_ops_dispatcher<data_parallel_policy, Float, Task, Method> {
                                   const descriptor_base<Task>& params,
                                   const train_input<Task>& input) const {
         using kernel_dispatcher_t =
-            dal::backend::kernel_dispatcher<backend::train_kernel_cpu<Float, Task, Method>,
-                                            backend::train_kernel_gpu<Float, Task, Method>>;
+            dal::backend::kernel_dispatcher<backend::train_kernel_cpu<Float, Method, Task>,
+                                            backend::train_kernel_gpu<Float, Method, Task>>;
         return kernel_dispatcher_t{}(ctx, params, input);
     }
 };
