+++ docs/source/onedal/algorithms/ensembles/includes/df-classification-examples.rst
@@ -20,7 +20,7 @@
 
     Batch Processing:
 
-    - :ref:`dpc_df_cls_dense_batch.cpp`
+    - :ref:`dpc_df_cls_hist_batch.cpp`
 
   .. group-tab:: oneAPI C++
