+++ docs/source/onedal/algorithms/ensembles/includes/df-regression-examples.rst
@@ -20,7 +20,7 @@
 
     Batch Processing:
 
-    - :ref:`dpc_df_reg_dense_batch.cpp`
+    - :ref:`dpc_df_reg_hist_batch.cpp`
 
   .. group-tab:: oneAPI C++
