+++ examples/oneapi/dpc/onedal_lnx.lst
@@ -19,8 +19,8 @@
 ##******************************************************************************
 
 ONEDAL = column_accessor_homogen           \
-         df_cls_dense_batch                \
-         df_reg_dense_batch                \
+         df_cls_hist_batch                 \
+         df_reg_hist_batch                 \
          kmeans_lloyd_dense_batch          \
          kmeans_init_dense                 \
          knn_cls_brute_force_dense_batch   \
